[package] 모델 패키지 분리
유저, 글, 댓글로 모델이 많아짐에 따라 해당 모델의 클래스명으로 패키지를 분리하여 각 클래스들을 이동하였습니다.
