+++ src/main/java/codesquad/http/filter/LoginCheckFilter.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import codesquad.http.dto.HttpRequest;
 import codesquad.http.status.HttpStatus;
 import codesquad.http.status.HttpStatusException;
-import codesquad.model.User;
+import codesquad.model.user.User;
 import codesquad.session.SessionManager;
 
 public class LoginCheckFilter implements Filter {
