+++ src/main/java/codesquad/model/user/UserRepository.java
@@ -1,0 +1,19 @@
+package codesquad.model.user;
+
+import java.util.List;
+
+import codesquad.annotation.Repository;
+
+@Repository
+public interface UserRepository {
+
+	User create(User user);
+
+	User findById(String userId);
+
+	User update(User user);
+
+	void delete(String userId);
+
+	List<User> findAll();
+}
