+++ src/main/java/codesquad/model/user/UserRepositoryImpl.java
@@ -1,4 +1,4 @@
-package codesquad.model;
+package codesquad.model.user;
 
 import java.util.ArrayList;
 import java.util.List;
